case-tecnico-3s/backend_django/requirements.txt
Django==4.2
graphene-django==3.2.1
requests==2.32.3
psycopg2-binary==2.9.9
pytest-django==4.8.0
django-cors-headers
